added order function

+++ src/main/java/sk/musicstore/repositories/DrumRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import sk.musicstore.models.Drum;
 
 @Repository
-public interface DrumRepository extends JpaRepository<Drum, Long> {
+public interface DrumRepository extends JpaRepository<Drum, Integer> {
 	
 }
