+++ src/main/java/sk/musicstore/repositories/SynthesizerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import sk.musicstore.models.Synthesizer;
 
 @Repository
-public interface SynthesizerRepository extends JpaRepository<Synthesizer, Long> {
+public interface SynthesizerRepository extends JpaRepository<Synthesizer, Integer> {
 
 }
